Refactor: View 인터페이스 단순화
과도하게 인터페이스를 분리하였다고 생각하여 기준을 세우고 적용하였습니다.
1. 테스트 할때 전략패턴이 필요한가?
2. 확장가능성이 있는가?
둘다 아니라고 생각하여 인터페이스를 모두 삭제했습니다.
