--- src/main/java/lotto/controller/view/InputView.java
@@ -1,10 +1,0 @@
-package lotto.controller.view;
-
-public interface InputView {
-
-    String inputAmountOfMoney();
-
-    String inputLottoNumber();
-
-    String inputBonusNumber();
-}
